Avoid abrupt end to expired session
Added a 30 min grace period to the end period for a circle session. This will let it run over by 30 min before its forced to end.
Fixed handling of the 'expiring' state to treat it as live state, before this fix it wasn't handled and shows a waiting room view.