Implementation of session timer
Implement `CountownTimer` that takes the following params:
- `startTime`: initial time to count from (i.e. when circle started)
- `endTime`: time when the countdown ends (i.e. circle.expiresOn)
- `displayValue`: whether to display the numeric time value as well as progress
- `color`: color of the countdown progress (and time value)
- `backgroundColor`: background of progress
- `displayType`: either hours:minutes or just minutes
- `valueLabel`: to display under time value
- `endValue`: to display when timer reaches 0
- `endValueLabel`: displayed under `endValue` when timer reaches 0
The timer runs every minute and updates the display. It handles changes to the colors, labels and the `endTime` The display of the numeric time value can be toggled by clicking on the widget.
Updated `ActiveSession` to keep the circle `startedOn` and `expiredOn` values. If `expiredOn` changes it will notify listeners
Update `CircleLiveVideoSession` to display a timer if the session has an expiration.  For keepers the timer shows the numeric value for the keeper but not members. When the session is about to expire the color changes as well as the display mode and the numeric value is shown for everyone. When the timer expires it shows "Now/ending".